add base dto
