--- CRUD/src/constants/entity.ts
@@ -1,6 +1,0 @@
-export enum ENTITY_NAME {
-  ORDER = "Order",
-  USER = "User",
-  PRODUCT = "Product",
-  CATEGORY = "Category",
-}
