+++ CRUD/src/constants/entityName.ts
@@ -1,0 +1,6 @@
+export enum ENTITY_NAME {
+  ORDER = "order",
+  USER = "user",
+  PRODUCT = "product",
+  CATEGORY = "category",
+}
