+++ CRUD/src/entities/category.entity.ts
@@ -1,5 +1,4 @@
-import { Column, Entity, ManyToMany, PrimaryGeneratedColumn } from "typeorm";
-import { Product } from "./product.entity";
+import { Column, Entity, PrimaryGeneratedColumn } from "typeorm";
 
 @Entity()
 export class Category {
@@ -8,7 +7,4 @@ export class Category {
 
   @Column({ unique: true })
   name: string;
-
-  @ManyToMany(() => Product, (product) => product.category)
-  products?: Product[];
 }
